abcall-pqrs-microservice/requirements.txt
PyDispatcher==2.0.6
boto3==1.35.19